Allow `property_value` tags in OBO files.
Fix the OboModel.pm module to prevent it from emitting a warning upon
reading a `property_value` tag. This tag is defined in the OBO Flatfile
format specification and should therefore be considered valid.

Treating this tag as "unrecognized" is especially problematic now that
we have switched metadata annotations from `oboInOwl:` to `dcterms:`,
because `dcterms:` annotations are unknown in the OBO Flatfile format
and therefore represented as generic `property_value` tags, instead of
specific tags such as `created_by` or `creation_date`.