Do not use the ontology-release-runner.
For historical reasons, FBcv release used to be generated by owltools'
ontology-release-runner (oort), completely bypassing most of the
standard ODK pipeline.

This is no longer warranted, and this commit removes most of the custom
pipeline to use the standard rules as much as possible. We still use a
custom rule to generate the -simple artifact (which is itself the basis
for the flybase_controlled_vocabulary.obo which is ultimately loaded
into CHADO). Our custom rule takes the standard main product ($(ONT))
and derives from it the -simple artifact almost exactly as the ODK would
do, but with an extra step to ensure the resulting product only contains
terms defined by FBcv).