Add 'driver-specific gene expression profile'.
Add a new type of dataset result for transcriptional profiles associated
with a specific transcriptional driver, when that driver is not known to
be associated with a precisely identified cell type.

This is primarily intended for TAPINseq-type datasets, where some
drivers may label unknown cell types (or possibly a mix of cell types).
Even without a cell type, downstream projects (especially VFB) might
still want the associated expression data.