Add 'split protein splicing tag'.
Add new term `split protein splicing tag` (will be needed to annotate
the CfaC and CfaN split inteins used in FBrf0259289) under `split gene
product engineering tool`.

We also amend the definition of `split system component` to remove the
bit stating that the reconstituted functional tool "would normally be
encoded by a single gene product", as it is not true for inteins.

closes #241