Derive fbcv.obo from the -simple artefact.
A comment in the custom Makefile says that we want fbcv.obo to be
derived from fbcv-simple, but the actual rule immediately below that
comment proceeds to derive fbcv.obo from fbcv.owl (which, as per the ODK
rules, is the same thing as fbcv-full.owl).

In all other FlyBase ontologies (FBdv, FBbt, and DPO), the main obo file
X.obo is derived from X-simple.owl, so there's no reason for FBcv to do
otherwise, especially since the comment explicitly states that this was
the original intention.