Add the mappings to the Alliance ontologies and vocabularies.
Add initial mappings to some of the ontologies and "vocabularies" used
at the Alliance. For now, this covers only the terms needed for
migrating the scRNAseq data.

The mappings to the "vocabularies" are kept in a separate set for
convenience, since they are "literal" mappings (vocabulary terms at the
Alliance are not proper semantic entities and do not have any kind of
public identifier).